moved get_local_host method to dockerless utils
