+++ packages/pytest-simcore/src/pytest_simcore/postgres_service.py
@@ -15,7 +15,7 @@ from servicelib.json_serialization import json_dumps
 from tenacity.stop import stop_after_delay
 from tenacity.wait import wait_fixed
 
-from .helpers.utils_docker import get_localhost_ip, get_service_published_port
+from .helpers.utils_host import get_localhost_ip, get_service_published_port
 from .helpers.utils_postgres import PostgresTestConfig, migrated_pg_tables_context
 
 _TEMPLATE_DB_TO_RESTORE = "template_simcore_db"
