+++ packages/pytest-simcore/src/pytest_simcore/redis_service.py
@@ -14,7 +14,7 @@ from tenacity.stop import stop_after_delay
 from tenacity.wait import wait_fixed
 from yarl import URL
 
-from .helpers.utils_docker import get_localhost_ip, get_service_published_port
+from .helpers.utils_host import get_localhost_ip, get_service_published_port
 
 log = logging.getLogger(__name__)
