+++ packages/pytest-simcore/src/pytest_simcore/simcore_services.py
@@ -20,7 +20,7 @@ from yarl import URL
 
 from .helpers.constants import MINUTE
 from .helpers.typing_env import EnvVarsDict
-from .helpers.utils_docker import get_localhost_ip, get_service_published_port
+from .helpers.utils_host import get_localhost_ip, get_service_published_port
 
 log = logging.getLogger(__name__)
