+++ services/osparc-gateway-server/tests/integration/test_clusters.py
@@ -11,7 +11,7 @@ from _pytest.monkeypatch import MonkeyPatch
 from aiodocker import Docker
 from dask_gateway import Gateway
 from faker import Faker
-from pytest_simcore.helpers.utils_docker import get_localhost_ip
+from pytest_simcore.helpers.utils_host import get_localhost_ip
 from tenacity._asyncio import AsyncRetrying
 from tenacity.stop import stop_after_delay
 from tenacity.wait import wait_fixed
