+++ services/osparc-gateway-server/tests/integration/test_dask_sidecar.py
@@ -7,7 +7,7 @@ from typing import Any, AsyncIterator, Awaitable, Callable
 import aiodocker
 import pytest
 from faker import Faker
-from pytest_simcore.helpers.utils_docker import get_localhost_ip
+from pytest_simcore.helpers.utils_host import get_localhost_ip
 from tenacity._asyncio import AsyncRetrying
 from tenacity.stop import stop_after_delay
 from tenacity.wait import wait_fixed
